backend/requirements.txt
attrs==24.2.0
blinker==1.8.2
click==8.1.7
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.1
PyJWT==2.9.0
pyodbc==5.1.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
